// frontend/src/types/MainGenres.ts
export enum MainGenres {
  RockName = 'Rock',
  RockKey = 'rock',
  PopName = 'Pop',
  PopKey = 'pop',
  HipHopName = 'Hip-Hop',
  HipHopKey = 'hip-hop',
  ElectronicName = 'Electronic',
  ElectronicKey = 'electronic',
  JazzName = 'Jazz',
  JazzKey = 'jazz',
  ClassicalName = 'Classical',
  ClassicalKey = 'classical',
};